requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
mpmath==1.2.1
psycopg2==2.9.3
pytz==2021.3
simpy==4.0.1
sqlparse==0.4.2
sympy==1.9
whitenoise==5.3.0
